+++ R/fit_aq_response.R
@@ -1,18 +1,15 @@
 #' Fit photosynthetic light-response curves
 #' 
-#' @inheritParams required_variables
-#' @param .data A data frame containing plant ecophysiological data. See \code{\link{required_variables()}}
-#' @param .vars A list to rename variables in .data to '.A' and '.Q' where '.A' is 
-#' net CO2 assimilation in \eqn{\mu}mol m\eqn{^{-2}} s\eqn{^{-1}} and '.Q' is 
-#' incident irradiance in \eqn{\mu}mol m\eqn{^{-2}} s\eqn{^{-1}}. '.Q' can be 
-#' corrected for light absorbance by setting `useapha_Q = TRUE` and setting 
-#' `alpha_Q = ...` based on measured or assumed leaf light absorbance. 
-#' @param .method A character string of the statistical method to use. Currently only "nls" (non-linear least squares) is allowed. We plan to add a Bayesion option later.
-#' @param usealpha_Q Flag. Should light intensity be multipled by `alpha_Q` before fitting? Default is FALSE (i.e. assume that 'Q' is absorbed light).
-#' @param alpha_Q Number. Absorbance of incident light. Default value is 0.84. Ignored if `usealpha_Q = FALSE`.
-#' @param brm_options A list of options passed to \code{\link[brms]{brm}} if `.method = "brms"`.
+#' @description We recommend using \code{\link{fit_photosynthesis}} with argument `.photo_fun = "aq_response"` rather than calling this function directly.
 #' 
-#' @return `fit_aq_response2()` fits the light response of net CO2 assimilation and returns an \code{\link[stats]{nls}} object.
+#' @inheritParams fit_photosynthesis
+#' @param usealpha_Q Flag. Should light intensity be multiplied by `alpha_Q` before fitting? Default is FALSE (i.e. assume that '.Q' is absorbed light).
+#' @param alpha_Q Number. Absorbance of incident light. Default value is 0.84. Ignored if `usealpha_Q = FALSE`.
+#' 
+#' @return 
+#' 
+#' * If `.method = 'ls'`: an \code{\link[stats]{nls}} object.
+#' * If `.method = 'brms'`: a \code{\link[brms]{brmsfit}} object.
 #' 
 #' @note Rd fitted in this way is essentially the same as the Kok method, and 
 #' represents a respiration value in the light that may not be accurate. 
@@ -42,9 +39,10 @@
 #'   mutate(group = as.factor(round(group, digits = -1)))
 #' 
 #' # Fit one light-response curve
-#' fit = fit_aq_response2(
+#' fit = fit_photosynthesis(
 #'   .data = filter(dat, group == 600),
-#'   .vars = list(.A = A, .Q = Qabs)
+#'   .photo_fun = "aq_response",
+#'   .vars = list(.A = A, .Q = Qabs),
 #' )
 #' 
 #' # The 'fit' object inherits class 'nls' and many methods can be used
@@ -67,39 +65,24 @@
 #' 
 #' fits = dat |>
 #'   split(~ group) |>
-#'   map(fit_aq_response2, .vars = list(.A = A, .Q = Qabs))
+#'   map(fit_photosynthesis, .photo_fun = "aq_response", .vars = list(.A = A, .Q = Qabs))
 #' 
 #' }
 #' 
 #' @md
 fit_aq_response2 = function(
     .data,
-    .vars = NULL,
     .model = "default",
-    .method = "nls",
+    .method = "ls",
     usealpha_Q = FALSE,
     alpha_Q = 0.84,
+    quiet = FALSE,
     brm_options = NULL
   ) {
   
-  .vars = substitute(.vars)
-  .method = match.arg(.method, choices = c("nls", "brms"))
-  
   # Checks
-  checkmate::assert_data_frame(.data)
-  checkmate::assert_string(.model)
   checkmate::assert_flag(usealpha_Q)
   checkmate::assert_number(alpha_Q, na.ok = !usealpha_Q)
-  
-  # Rename variables
-  if (!is.null(.vars)) {
-    deparse(.vars) |>
-      stringr::str_replace("^list\\(", ".data = dplyr::rename(.data, ") |>
-      str2lang() |>
-      eval()
-  }
-  
-  checkmate::assert_subset(c(".A", ".Q"), choices = colnames(.data))
   
   # Set light intensity dependent on whether it is incident or
   # absorbed that you want the variables on
@@ -109,7 +92,7 @@ fit_aq_response2 = function(
   # Fit AQ response model
   fit = switch(
     .method,
-    nls = fit_aq_response2_nls(.data),
+    ls = fit_aq_response2_nls(.data),
     brms = fit_aq_response2_brms(.data, brm_options)
   )
  
@@ -306,7 +289,11 @@ fit_aq_response = function(
     title = NULL
 ) {
   
-  lifecycle::deprecate_warn("2.1.1", "fit_aq_response()", "fit_aq_response2()", always = TRUE)
+  lifecycle::deprecate_warn(
+    "2.1.1", "fit_aq_response()", 
+    "fit_photosynthesis(.photo_fun = 'aq_response')", 
+    always = FALSE
+  )
   
   # Locally bind variables - avoids notes on check package
   A_net = NULL