+++ tests/testthat/test-fit_aq_response2.R
@@ -1,7 +1,7 @@
-context("fit_aq_response2()")
+context("fit_photosynthesis()")
 library(photosynthesis)
 
-test_that("'fit_aq_response2()' accepts data.frames and tibbles", {
+test_that("'fit_photosynthesis()' accepts data.frames and tibbles", {
   
   df1 = data.frame(
     .A = c(10, 9.5, 8, 3.5, 2.5, 2.0, 1, 0.2),
@@ -9,8 +9,8 @@ test_that("'fit_aq_response2()' accepts data.frames and tibbles", {
   )
   df2 = tibble::as_tibble(df1)
   
-  expect_no_condition({fit_aq_response2(.data = df1)})
-  expect_no_condition({fit_aq_response2(.data = df2)})
+  expect_no_condition({fit_photosynthesis(.data = df1, .photo_fun = "aq_response")})
+  expect_no_condition({fit_photosynthesis(.data = df2, .photo_fun = "aq_response")})
   
 })
 
@@ -24,25 +24,25 @@ test_that(".vars argument renames variables", {
   df3 = dplyr::rename(df1, PPFD = .Q)
   df4 = dplyr::rename(df1, Photo = .A, PPFD = .Q)
   
-  expect_error({fit_aq_response2(.data = df2)})
-  expect_error({fit_aq_response2(.data = df3)})
-  expect_error({fit_aq_response2(.data = df4)})
+  expect_error({fit_photosynthesis(.data = df2, .photo_fun = "aq_response")})
+  expect_error({fit_photosynthesis(.data = df3, .photo_fun = "aq_response")})
+  expect_error({fit_photosynthesis(.data = df4, .photo_fun = "aq_response")})
 
   expect_no_condition({
-    fit_aq_response2(.data = df2, .vars = list(.A = Photo))
+    fit_photosynthesis(.data = df2, .photo_fun = "aq_response", .vars = list(.A = Photo))
   })
   expect_no_condition({
-    fit_aq_response2(.data = df3, .vars = list(.Q = PPFD))
+    fit_photosynthesis(.data = df3, .photo_fun = "aq_response", .vars = list(.Q = PPFD))
   })
   expect_no_condition({
-    fit_aq_response2(.data = df4, .vars = list(.A = Photo, .Q = PPFD))
+    fit_photosynthesis(.data = df4, .photo_fun = "aq_response", .vars = list(.A = Photo, .Q = PPFD))
   })
   
-  expect_error({fit_aq_response2(.data = df2, .vars(.A = foo))})
-  expect_error({fit_aq_response2(.data = df3, .vars(.Q = bar))})
+  expect_error({fit_photosynthesis(.data = df2, .photo_fun = "aq_response", .vars(.A = foo))})
+  expect_error({fit_photosynthesis(.data = df3, .photo_fun = "aq_response", .vars(.Q = bar))})
 
   expect_no_condition({
-    fit_aq_response2(.data = dplyr::mutate(df4, A1 = Photo), 
+    fit_photosynthesis(.data = dplyr::mutate(df4, A1 = Photo), .photo_fun = "aq_response",
                      .vars = list(.A = Photo, .Q = PPFD))
   })